split into index and browser
