+++ browser.js
@@ -1,0 +1,3 @@
+const { createClient } = require('./lib/client_factory')
+
+exports.createClient = createClient
