+++ lib/client_identity.js
