+++ package.json
@@ -3,11 +3,13 @@
   "description": "A HTTP client working by the Module Standard",
   "version": "1.1.1",
   "main": "index.js",
+  "browser": "browser.js",
   "repository": "https://github.com/FirstLegoLeague/ms-client.git",
   "author": "Idan Stark <idanstark42@gmail.com>",
   "license": "GPL-3.0",
   "files": [
     "index.js",
+    "browser.js",
     "lib/**/*.js"
   ],
   "engines": {
@@ -46,7 +48,6 @@
     "axios": "^0.19.0"
   },
   "dependencies": {
-    "detect-node": "^2.0.4",
     "randomatic": "^3.1.1"
   },
   "optionalDependencies": {
