+++ test/lib/client_identity.test.js
